[clang][deps] Consolidate types into new `DependencyActionController.h`
