+++ clang/lib/DependencyScanning/DependencyScanningWorker.cpp
@@ -38,7 +38,6 @@ DependencyScanningWorker::DependencyScanningWorker(
 }
 
 DependencyScanningWorker::~DependencyScanningWorker() = default;
-DependencyActionController::~DependencyActionController() = default;
 
 static bool createAndRunToolInvocation(
     ArrayRef<std::string> CommandLine, DependencyScanningAction &Action,
